+++ src/schemathesis/cli/output.py
@@ -1,10 +1,17 @@
+import os
+import platform
 import shutil
 from contextlib import contextmanager
 from typing import Generator, List, Optional
 
 import click
+from hypothesis import settings
+from importlib_metadata import version
 
 from .. import runner
+from ..constants import __version__
+from ..models import StatsCollector
+from ..runner import events
 
 
 @contextmanager
@@ -62,3 +69,65 @@ def pretty_print_stats(stats: runner.StatsCollector, hypothesis_output: Optional
                     click.style(verdict, fg=color, bold=True),
                 )
             )
+
+
+def percentage(position: int, length: int) -> str:
+    return f"[{position * 100 // length}%]"
+
+
+def pretty_print_test_progress(results_generator: Generator[events.ExecutionEvent, None, None]) -> StatsCollector:
+    # Starting progress position
+    position = 0
+    errors = 0
+    # Terminal size and fields position
+    columns = shutil.get_terminal_size((100, 50)).columns
+    col1_len = 8  # Method column
+    col2_len = 10  # Test endpoint column
+    col3_len = 4  # Test result column
+    col4_len = 10  # Progress column
+    # Initial runner state
+    init = next(results_generator)
+    print_header(init.schema.endpoints_count)  # type: ignore
+    for result in results_generator:
+        # Print test method and endpoint before test execution
+        if isinstance(result, events.BeforeExecution):
+            # Collect errors count before test execution
+            errors = sum([check["error"] for check in result.statistic.data.values()])
+            # Re-calculate fields position
+            col2_len = len(result.endpoint.path)
+            template = f"    {{:<{col1_len}}} {{:<{col2_len}}} "
+            click.echo(template.format(result.endpoint.method, result.endpoint.path), nl=False)
+        # Print test result and progress after execution
+        if isinstance(result, events.AfterExecution):
+            position += 1
+            # Recalculate number of errors to obtain test state after execution
+            new_errors = sum([check["error"] for check in result.statistic.data.values()])
+            template = f"{{:{col3_len}}} "
+            if new_errors > errors:
+                click.echo(template.format("F"), nl=False)
+            else:
+                click.echo(template.format("."), nl=False)
+            col4_len = int(columns) - col1_len - col2_len - col3_len - 10
+            click.echo(f"{{:>{col4_len}}}".format(percentage(position, init.schema.endpoints_count)))  # type: ignore
+        if isinstance(result, events.Finished):
+            return result.statistic
+    return StatsCollector()
+
+
+def print_header(tests_number: int) -> None:
+    with print_in_section("test session starts"):
+        versions = (
+            f"platform {platform.system()} -- "
+            f"Python {platform.python_version()}, "
+            f"schemathesis-{__version__}, "
+            f"hypothesis-{version('hypothesis')}, "
+            f"hypothesis_jsonschema-{version('hypothesis_jsonschema')}, "
+            f"jsonschema-{version('jsonschema')}"
+        )
+        click.echo(versions)
+        click.echo(f"rootdir: {os.getcwd()}")
+        click.echo(
+            f"hypothesis profile '{settings._current_profile}' "  # type: ignore
+            f"-> {settings.get_profile(settings._current_profile).show_changed()}"
+        )
+        click.echo(f"Collected endpoints: {tests_number}")